Rename sbi_files.{c,h} -> filesystem.{c,h}
