+++ machine/filesystem.c
@@ -1,4 +1,4 @@
-#include "sbi_files.h"
+#include "filesystem.h"
 #include "tinycstd.h"
 
 const KFILE* find_file(const char* filename) {
